requirements.txt
appdirs==1.4.4
psutil==5.9.0
pystray==0.19.3
python-dateutil==2.8.2
watchdog==2.1.7
